+++ packages/charts/src/utils/point.ts
@@ -12,7 +12,7 @@ export interface Point {
   y: number;
 }
 
-/** @internal * */
+/** @internal */
 export function getDelta(start: Point, end: Point) {
   return Math.sqrt(Math.pow(end.x - start.x, 2) + Math.pow(end.y - start.y, 2));
 }