+++ storybook/stories/interactions/16_cursor_update_action.story.tsx
@@ -27,7 +27,7 @@ import {
   LegendValue,
 } from '@elastic/charts';
 import { KIBANA_METRICS } from '@elastic/charts/src/utils/data_samples/test_dataset_kibana';
-import { palettes } from '@elastic/charts/src/utils/themes/colors';
+import { palettes } from '@elastic/charts/src/utils/themes/legacy_colors';
 
 import type { ChartsStory } from '../../types';
 import { useBaseTheme } from '../../use_base_theme';